Adds CoursePolicy and checks.

+++ app/Http/Controllers/CoursesController.php
@@ -68,6 +68,8 @@ class CoursesController extends Controller {
      */
     public function edit(Course $course)
     {
+        $this->authorize('updateCourse', $course);
+
         return view('course.edit', ['course' => $course]);
     }
 
@@ -80,6 +82,8 @@ class CoursesController extends Controller {
      */
     public function update(CourseRequest $request, Course $course)
     {
+        $this->authorize('updateCourse', $course);
+
         $course->update($request->all());
 
         return redirect()->action('CoursesController@show', ['course' => $course]);
@@ -94,6 +98,8 @@ class CoursesController extends Controller {
      */
     public function destroy(Course $course)
     {
+        $this->authorize('updateCourse', $course);
+
         $course->delete();
 
         return redirect()->action('CoursesController@index');
